:white_check_mark: Add tests for readme resources

+++ tests/test_app.py
@@ -1,5 +1,5 @@
 from kf_model_fhir import loader, cli
-from conftest import TEST_DATA_DIR, PROFILE_DIR, RESOURCE_DIR
+from conftest import PROFILE_DIR, RESOURCE_DIR
 import os
 import shutil
 
@@ -49,7 +49,17 @@ def test_fhir_format(caplog, tmpdir, path, expected_exc):
     os.remove(out_file)
 
 
-def test_add_resource_to_ig(temp_ig):
+@pytest.mark.parametrize(
+    "dir_list, expected_code",
+    [
+        ([os.path.join(PROFILE_DIR, 'valid'),
+          os.path.join(RESOURCE_DIR, 'valid'),
+          os.path.join(PROFILE_DIR, 'extensions')], 0),
+        ([os.path.join(PROFILE_DIR, 'invalid'),
+          os.path.join(RESOURCE_DIR, 'invalid')], 1),
+    ],
+)
+def test_ig_validation(temp_ig, dir_list, expected_code):
     """
     Test kf_model_fhir.app.add_resource_to_ig
     """
@@ -58,21 +68,22 @@ def test_add_resource_to_ig(temp_ig):
     temp_ig_control_file = os.path.join(temp_site_root, 'ig.json')
 
     # Add conformance resource and example resources to IG
-    files = ['profiles/extensions/StructureDefinition-proband-status.json',
-             'profiles/StructureDefinition-Participant.json',
-             'resources/participant-example.json']
-    for f in files:
-        source = os.path.join(TEST_DATA_DIR, f)
+    filepaths = [os.path.join(d, f) for d in dir_list for f in os.listdir(d)]
+    for source in filepaths:
+        if not os.path.isfile(source):
+            continue
+        filename = os.path.split(source)[-1]
         opts = ['--ig_control_file', temp_ig_control_file,
                 '--is_example']
-        filename = os.path.split(f)[-1]
-        if f.startswith('profiles'):
+        if 'profiles' in os.path.split(source)[0]:
             dest = os.path.join(
                 temp_site_root, 'source', 'resources', filename
             )
             opts.pop()
         else:
             dest = os.path.join(temp_site_root, 'source', 'examples', filename)
+
+        # Copy test file into temp IG
         shutil.copyfile(source, dest)
         result = runner.invoke(cli.add, [dest] + opts)
 
@@ -84,45 +95,10 @@ def test_add_resource_to_ig(temp_ig):
         cli.validate,
         [temp_ig_control_file, '--publisher_opts', '-tx n/a']
     )
-    assert result.exit_code == 0
+    assert result.exit_code == expected_code
 
 
-def test_invalid_ig(temp_ig):
-    """
-    Make sure invalid IG fails validation
-    """
-    runner = CliRunner()
-    temp_site_root = temp_ig
-    temp_ig_control_file = os.path.join(temp_site_root, 'ig.json')
-
-    # Add invalid conformance resources
-    invalid_profile_dir = os.path.join(PROFILE_DIR, 'invalid')
-    for fn in os.listdir(invalid_profile_dir):
-        src = os.path.join(PROFILE_DIR, 'invalid', fn)
-        dst = os.path.join(temp_site_root, 'source', 'resources', fn)
-        shutil.copyfile(src, dst)
-        result = runner.invoke(
-            cli.add, [dst, '--ig_control_file', temp_ig_control_file]
-        )
-        assert result.exit_code == 0
-
-    # Add invalid example resources
-    invalid_example_dir = os.path.join(RESOURCE_DIR, 'invalid')
-    if os.path.exists(invalid_example_dir):
-        for fn in os.listdir(invalid_profile_dir):
-            src = os.path.join(PROFILE_DIR, 'invalid', fn)
-            dst = os.path.join(temp_site_root, 'source', 'examples', fn)
-            shutil.copyfile(src, dst)
-            result = runner.invoke(
-                cli.add,
-                [dst, '--is_example',
-                 '--ig_control_file', temp_ig_control_file]
-            )
-            assert result.exit_code == 0
-
-    # Validate IG
-    result = runner.invoke(
-        cli.validate,
-        [temp_ig_control_file, '--publisher_opts', '-tx n/a']
-    )
-    assert result.exit_code != 0
+if __name__ == '__main__':
+    test_ig_validation('boo', [os.path.join(PROFILE_DIR, 'valid'),
+                               os.path.join(RESOURCE_DIR, 'valid'),
+                               os.path.join(PROFILE_DIR, 'extensions')], 0)
